core/kohya_inference_manager.py

# core/kohya_inference_manager.py
"""
Unified Inference Manager leveraging Kohya's library for image generation.
"""
import os
import sys
import torch
from typing import Dict, Any, Optional

# Add Kohya's backend to system path
sys.path.insert(0, os.path.join(os.getcwd(), "trainer", "derrian_backend", "sd_scripts"))

from library import model_util, train_util
from library.strategy_sd import SdTokenizeStrategy, SdTextEncodingStrategy
from library.strategy_sdxl import SdxlTokenizeStrategy, SdxlTextEncodingStrategy
# Import other strategies as needed, e.g., flux, sd3

import logging
logging.basicConfig(level=logging.INFO)
logger = logging.getLogger(__name__)

class KohyaInferenceManager:
    def __init__(self):
        self.strategies = {
            'sd15': {'tokenize': SdTokenizeStrategy(v2=False, max_length=75), 'encoding': SdTextEncodingStrategy(clip_skip=1)},
            'sdxl': {'tokenize': SdxlTokenizeStrategy(max_length=75), 'encoding': SdxlTextEncodingStrategy()},
            # 'flux': {'tokenize': FluxTokenizeStrategy, 'encoding': FluxTextEncodingStrategy},
            # 'sd3': {'tokenize': Sd3TokenizeStrategy, 'encoding': Sd3TextEncodingStrategy}
        }
        self.models_cache = {}

    def detect_model_type(self, model_path: str) -> str:
        """
        Detects the model type (e.g., 'sd15', 'sdxl') from the model path.
        """
        model_path_lower = model_path.lower()
        if 'sdxl' in model_path_lower or 'xl' in model_path_lower:
            return 'sdxl'
        # Add more sophisticated detection logic as needed
        return 'sd15'

    def _load_model(self, model_path: str, model_type: str):
        """
        Loads the model, text_encoder, and VAE from a checkpoint.
        Caches the loaded models to avoid reloading.
        """
        if model_path in self.models_cache:
            return self.models_cache[model_path]

        logger.info(f"Loading {model_type} model from: {model_path}")
        
        v2 = 'v2' in model_path.lower()
        
        text_encoder, vae, unet = model_util.load_models_from_stable_diffusion_checkpoint(
            v2=v2,
            ckpt_path=model_path
        )
        
        self.models_cache[model_path] = (text_encoder, vae, unet)
        return text_encoder, vae, unet

    def generate_image(self, params: Dict[str, Any]):
        """
        Generates an image based on the provided parameters.
        """
        model_path = params.get('model_path')
        prompt = params.get('prompt', '')
        
        if not model_path or not os.path.exists(model_path):
            logger.error("Model path is required and must exist.")
            return

        model_type = self.detect_model_type(model_path)
        if model_type not in self.strategies:
            logger.error(f"Unsupported model type: {model_type}")
            return

        strategy = self.strategies[model_type]
        
        try:
            text_encoder, vae, unet = self._load_model(model_path, model_type)
            
            # TODO: Implement the full inference pipeline:
            # 1. Tokenize prompt using strategy['tokenize']
            # 2. Encode tokens using strategy['encoding']
            # 3. Prepare latents (noise)
            # 4. Denoising loop using the unet
            # 5. Decode latents using VAE
            # 6. Save or return the image
            
            logger.info(f"Successfully loaded model and strategy for {model_type}.")
            logger.info("Inference pipeline is not fully implemented yet.")

        except Exception as e:
            logger.error(f"An error occurred during image generation: {e}")
            # Clean up cache if loading failed
            if model_path in self.models_cache:
                del self.models_cache[model_path]

# Example usage:
if __name__ == '__main__':
    manager = KohyaInferenceManager()
    # This is a placeholder for a real model path
    sd15_model = '/path/to/your/sd15.safetensors'
    sdxl_model = '/path/to/your/sdxl.safetensors'

    if os.path.exists(sd15_model):
        manager.generate_image({
            'model_path': sd15_model,
            'prompt': 'a beautiful landscape painting'
        })
    
    if os.path.exists(sdxl_model):
        manager.generate_image({
            'model_path': sdxl_model,
            'prompt': 'an astronaut riding a horse on mars'
        })
