requirements-backend.txt
# ===================================================================
# Unified Backend Requirements for LoRA Easy Training Jupyter
#
# This file is auto-generated and managed by the SetupManager.
# It consolidates dependencies from all submodules.
# ===================================================================

# -- Core Training Dependencies (from Kohya sd-scripts) --
accelerate==0.33.0
transformers==4.44.0
diffusers[torch]==0.25.0
safetensors==0.4.4
bitsandbytes==0.44.0
torch-fidelity==0.3.0
# Note: torch, torchvision, and torchaudio are installed by the setup manager
# based on the detected GPU environment (CUDA, ROCm, etc.)

# -- General Utilities --
toml==0.10.2
einops==0.8.1
tqdm==4.66.4
rich==13.7.1
ftfy==6.2.0
requests>=2.28.0
psutil
pyyaml
pandas
scipy
omegaconf
huggingface-hub>=0.10.0
fiftyone
gallery-dl
hf-transfer
voluptuous==0.15.2
sentencepiece==0.2.0

# -- Image/UI/Data Handling --
Pillow<10.0.0
opencv-python==4.8.1.78
imageio==2.34.1
imageio-ffmpeg==0.5.1
pyarrow==16.1.0
altair==5.3.0
streamlit==1.35.0
easygui==0.98.3
Jinja2==3.1.4
sidecar>=0.5.0
ipywidgets>=8.0.0
ipython>=8.0.0
imagesize==1.4.1

# -- Optimizers & Schedulers --
prodigyopt==1.0
lion-pytorch==0.1.2
dadaptation==3.2
pytorch_optimizer>=3.1.0
schedulefree
prodigy-plus-schedule-free
came-pytorch

# -- Monitoring & Logging --
tensorboard==2.16.2
wandb==0.17.0
watchdog==4.0.0

# -- ONNX / Tagger Dependencies --
onnx==1.16.1
onnxruntime-gpu==1.18.0
protobuf==4.25.3

# -- Windows Specific (safe to include for all OS) --
pyreadline
colorama

# -- Deprecated/Unused from old files (for reference) --
# distro (Linux distro detection - unused)
# pytorch-lightning (only used in LyCORIS experiments)
